+++ src/PhpSpreadsheet/Shared/Escher/DggContainer.php
@@ -2,6 +2,8 @@
 
 namespace PhpOffice\PhpSpreadsheet\Shared\Escher;
 
+use PhpOffice\PhpSpreadsheet\Exception as SpreadsheetException;
+
 class DggContainer
 {
     /**
@@ -92,6 +94,14 @@ class DggContainer
     public function getBstoreContainer(): ?DggContainer\BstoreContainer
     {
         return $this->bstoreContainer;
+    }
+
+    /**
+     * Get BLIP Store Container.
+     */
+    public function getBstoreContainerOrThrow(): DggContainer\BstoreContainer
+    {
+        return $this->bstoreContainer ?? throw new SpreadsheetException('bstoreContainer is unexpectedly null');
     }
 
     /**